@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Device.Gpio;
+﻿using System.Device.Gpio;
 
 namespace FlushTheToiletWebServer.CF01.Devices
 {
